let flask put server ip in script
Moving the javascript into the index.html file lets the template
thing put in the request's IP address (instead of editing the
javascript). The tank runs either as a client on the home wifi, or
as a wifi ap when in the wild (with a different IP in that case).

Also adjust motor speeds so the tank goes faster by default.